pyproject.toml
[tool.poetry]
name = "drivingschool-metrics"
version = "0.1.0"
description = ""
authors = ["josericardopenase <josericardopenase@gmail.com>"]
readme = "README.md"
packages = [{include = "config"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
